Changed getter method for size of class DataSet to return a size_t, added a placeholder for a class to manage parsing of command line arguments, Added tests to unit test of class DataSet to test, if after copy construction, neither Events nor their internal arrays point to the same memory

+++ src/DataSet.cpp
@@ -130,7 +130,7 @@ void DataSet::addData(std::unique_ptr<Event> data)
  *
  * @ensure size >= 0
  */
-unsigned int DataSet::getSize() const
+size_t DataSet::getSize() const
 {
 	return m_data.size();
 }
